refactor : repository interface를 분리, model과 함께 domain 영역으로 이동
