+++ src/main/java/kitchenpos/menu/dao/JdbcTemplateMenuProductDao.java
@@ -1,6 +1,7 @@
 package kitchenpos.menu.dao;
 
-import kitchenpos.menu.model.MenuProduct;
+import kitchenpos.menu.domain.MenuProductDao;
+import kitchenpos.menu.domain.model.MenuProduct;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
