+++ src/main/java/kitchenpos/menu/domain/MenuDao.java
@@ -1,6 +1,6 @@
-package kitchenpos.menu.dao;
+package kitchenpos.menu.domain;
 
-import kitchenpos.menu.model.Menu;
+import kitchenpos.menu.domain.model.Menu;
 
 import java.util.List;
 import java.util.Optional;
