+++ src/main/java/kitchenpos/order/dao/JdbcTemplateOrderLineItemDao.java
@@ -1,6 +1,7 @@
 package kitchenpos.order.dao;
 
-import kitchenpos.order.model.OrderLineItem;
+import kitchenpos.order.domain.OrderLineItemDao;
+import kitchenpos.order.domain.model.OrderLineItem;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
