+++ src/main/java/kitchenpos/order/domain/OrderLineItemDao.java
@@ -1,6 +1,6 @@
-package kitchenpos.order.dao;
+package kitchenpos.order.domain;
 
-import kitchenpos.order.model.OrderLineItem;
+import kitchenpos.order.domain.model.OrderLineItem;
 
 import java.util.List;
 import java.util.Optional;
