+++ src/main/java/kitchenpos/order/domain/model/Order.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.model;
+package kitchenpos.order.domain.model;
 
 import java.time.LocalDateTime;
 import java.util.List;
