+++ src/main/java/kitchenpos/order/domain/model/OrderLineItem.java
@@ -1,9 +1,9 @@
-package kitchenpos.menu.model;
+package kitchenpos.order.domain.model;
 
-public class MenuProduct {
+public class OrderLineItem {
     private Long seq;
+    private Long orderId;
     private Long menuId;
-    private Long productId;
     private long quantity;
 
     public Long getSeq() {
@@ -14,20 +14,20 @@ public class MenuProduct {
         this.seq = seq;
     }
 
+    public Long getOrderId() {
+        return orderId;
+    }
+
+    public void setOrderId(final Long orderId) {
+        this.orderId = orderId;
+    }
+
     public Long getMenuId() {
         return menuId;
     }
 
     public void setMenuId(final Long menuId) {
         this.menuId = menuId;
-    }
-
-    public Long getProductId() {
-        return productId;
-    }
-
-    public void setProductId(final Long productId) {
-        this.productId = productId;
     }
 
     public long getQuantity() {
