+++ src/main/java/kitchenpos/order/domain/model/OrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.model;
+package kitchenpos.order.domain.model;
 
 public enum OrderStatus {
     COOKING, MEAL, COMPLETION
