+++ src/main/java/kitchenpos/orderTable/service/OrderTableService.java
@@ -1,9 +1,9 @@
 package kitchenpos.orderTable.service;
 
-import kitchenpos.order.dao.OrderDao;
-import kitchenpos.order.model.OrderStatus;
-import kitchenpos.orderTable.dao.OrderTableDao;
-import kitchenpos.orderTable.model.OrderTable;
+import kitchenpos.order.domain.OrderDao;
+import kitchenpos.order.domain.model.OrderStatus;
+import kitchenpos.orderTable.domain.OrderTableDao;
+import kitchenpos.orderTable.domain.model.OrderTable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
