+++ src/main/java/kitchenpos/product/dao/JdbcTemplateProductDao.java
@@ -1,6 +1,7 @@
 package kitchenpos.product.dao;
 
-import kitchenpos.product.model.Product;
+import kitchenpos.product.domain.ProductDao;
+import kitchenpos.product.domain.model.Product;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
