+++ src/test/java/kitchenpos/order/service/OrderServiceTest.java
@@ -1,15 +1,15 @@
 package kitchenpos.order.service;
 
 import kitchenpos.menu.dao.InMemoryMenuDao;
-import kitchenpos.menu.dao.MenuDao;
+import kitchenpos.menu.domain.MenuDao;
 import kitchenpos.order.dao.InMemoryOrderDao;
 import kitchenpos.order.dao.InMemoryOrderLineItemDao;
-import kitchenpos.order.dao.OrderDao;
-import kitchenpos.order.dao.OrderLineItemDao;
-import kitchenpos.order.model.Order;
-import kitchenpos.order.model.OrderStatus;
+import kitchenpos.order.domain.OrderDao;
+import kitchenpos.order.domain.OrderLineItemDao;
+import kitchenpos.order.domain.model.Order;
+import kitchenpos.order.domain.model.OrderStatus;
 import kitchenpos.orderTable.dao.InMemoryOrderTableDao;
-import kitchenpos.orderTable.dao.OrderTableDao;
+import kitchenpos.orderTable.domain.OrderTableDao;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
