+++ src/test/java/kitchenpos/orderTable/controller/TableRestControllerTest.java
@@ -1,6 +1,6 @@
 package kitchenpos.orderTable.controller;
 
-import kitchenpos.orderTable.model.OrderTable;
+import kitchenpos.orderTable.domain.model.OrderTable;
 import kitchenpos.orderTable.service.OrderTableService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
