+++ src/test/java/kitchenpos/orderTable/dao/InMemoryTableGroupDao.java
@@ -1,6 +1,7 @@
 package kitchenpos.orderTable.dao;
 
-import kitchenpos.orderTable.model.TableGroup;
+import kitchenpos.orderTable.domain.TableGroupDao;
+import kitchenpos.orderTable.domain.model.TableGroup;
 
 import java.util.*;
