+++ src/test/java/kitchenpos/product/controller/ProductRestControllerTest.java
@@ -1,6 +1,6 @@
 package kitchenpos.product.controller;
 
-import kitchenpos.product.model.Product;
+import kitchenpos.product.domain.model.Product;
 import kitchenpos.product.service.ProductService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
